First attempt at a display function (for html syntax) that uses the output of an incremental lexer and parser.  This code is not complete yet:
  * right now, it uses only the lexer output, and not the parser

  * the cursor is not displayed yet

  * it is too slow, most likely because the output records are all in
    one big bag, as opposed to being tree structured, such as the
    lines of basic syntax.

The slowness is not a serious problem, because one day, McCLIM will
have tree-structured output records itself, and because most syntax
modules (including this one, very soon) will have some tree structure
itself.  It might be worthwhile to display the part of the buffer
beyond a parse error in some artificially structured way, such as by
lines as in the current basic syntax.