Experimental fix for CSR minibuffer problems.
Oddly, not everyone suffers from this.  Nor from the problem the
handle-repaint method was included to solve: the scrollbars not
adjusting properly.

handle-repaint is not necessarily the right place for this functionality
to be hooked on, but note-sheet-region-changed isn't either: drawing to
a viewport also seems to change the region, so drawing to the frame from
within the note-sheed-region-changed is a bad idea.