New syntax: `fundamental'
My idea is to eventually get rid of the `basic' syntax.  Not only is
the name not great, but the code it contains some cruft that I would
like to get rid of.  For instance, the cache is no longer necessary
and is messy to manage anyway.

In addition, the `basic' syntax is problematic with respect to factoring
out the GUI part of Climacs into a separate application-independent
module.  That's another good reason to get rid of it.

So, I would appreciate if you would please check out this new syntax
and let me know how it turns out.  If I hear no complaints, I'll start
by making `fundamental' the default syntax, and then remove the
`basic' syntax altogether.