Removed functions find-climacs-pane and find-real-pane because they are no longer needed.
Removed stupid names from panes, because they are not needed.

Wrote a new version of set-key that can take a list of key strokes and
that creates nested command tables as needed.  Modified the esa
example to take advantage of this new feature.  Now, Climacs itself
should probably be modified to take advantage of it.