Implemented `shallow command tables'.
Made the ESA command loop search for key bindings in the inherit-from
list as well.

Changed the ESA example so that com-quit is in the esa-global-table
and the example-global-table inherits from the esa-global-table.

Next, it would be good to create many small command tables that
contain (say) all the commands that have to do with multi-windowing
(C-x 2, C-x 3, etc), all the commands that have to do with kbd macros,
all the commands that have to do with undo, etc.

Also, next, rearrange Climacs itself to take advantage of all this.